requirements.txt
asgiref==3.11.0
bcrypt==5.0.0
Django==5.1.5
django-extensions==4.1
djangorestframework==3.16.1
pip==24.0
PyJWT==2.10.1
python-dotenv==1.2.1
sqlparse==0.5.5
psycopg2-binary==2.9.11